requirements.txt
Django==2.2.4
django-polls==0.1
django-polymorphic==2.1.2
djangorestframework==3.10.3
pkg-resources==0.0.0
pytz==2019.2
sqlparse==0.3.0
